+++ profiler/src/test/java/com/splunk/opentelemetry/profiler/JfrSettingsOverridesTest.java
@@ -16,12 +16,15 @@
 
 package com.splunk.opentelemetry.profiler;
 
+import static com.splunk.opentelemetry.profiler.Configuration.CONFIG_KEY_CALL_STACK_INTERVAL;
+import static com.splunk.opentelemetry.profiler.Configuration.CONFIG_KEY_DEPRECATED_THREADDUMP_PERIOD;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotSame;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import io.opentelemetry.instrumentation.api.config.Config;
+import java.time.Duration;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 
@@ -30,23 +33,35 @@ class JfrSettingsOverridesTest {
   @Test
   void testOverrides() {
     Config config = mock(Config.class);
-    when(config.getString("splunk.profiler.period.threaddump")).thenReturn("163");
-    when(config.getString("splunk.profiler.period.otherevent")).thenReturn("964");
-    when(config.getString("splunk.profiler.period.extraunused")).thenReturn("111");
+    when(config.getDuration(CONFIG_KEY_CALL_STACK_INTERVAL, Duration.ZERO))
+        .thenReturn(Duration.ofMillis(163));
     when(config.getBoolean("splunk.profiler.tlab.enabled", false)).thenReturn(true);
     JfrSettingsOverrides overrides = new JfrSettingsOverrides(config);
     Map<String, String> jfrSettings =
         Map.of(
             "jdk.ThreadDump#period", "12",
             "jdk.ThreadDump#enabled", "true",
-            "jdk.OtherEvent#period", "13",
             "jdk.ObjectAllocationInNewTLAB#enabled", "true",
             "jdk.ObjectAllocationOutsideTLAB#enabled", "true");
     Map<String, String> result = overrides.apply(jfrSettings);
     assertNotSame(result, jfrSettings);
     assertEquals("163 ms", result.get("jdk.ThreadDump#period"));
-    assertEquals("964 ms", result.get("jdk.OtherEvent#period"));
     assertEquals("true", result.get("jdk.ObjectAllocationInNewTLAB#enabled"));
     assertEquals("true", result.get("jdk.ObjectAllocationOutsideTLAB#enabled"));
+  }
+
+  @Test
+  void fallBackToDeprecatedConfig() {
+    Config config = mock(Config.class);
+    when(config.getDuration(CONFIG_KEY_DEPRECATED_THREADDUMP_PERIOD, Duration.ZERO))
+        .thenReturn(Duration.ofMillis(999));
+    JfrSettingsOverrides overrides = new JfrSettingsOverrides(config);
+    Map<String, String> jfrSettings =
+        Map.of(
+            "jdk.ThreadDump#period", "12",
+            "jdk.ThreadDump#enabled", "true");
+    Map<String, String> result = overrides.apply(jfrSettings);
+    assertNotSame(result, jfrSettings);
+    assertEquals("999 ms", result.get("jdk.ThreadDump#period"));
   }
 }